Updated color schemes... mostly works
Also, now the alarm handler tab is "expert" tab, and it
now allows user editting without pausing the loop (so
watch out).

Now I need to make sure that alarm status gets set properly
and that it will set the local and global alarm latches,
and that it is read properly, and that user silencing will
actually silence.